+++ src/main/java/io/kontur/eventapi/normalization/Normalizer.java
@@ -46,4 +46,8 @@ public abstract class Normalizer implements Applicable<DataLake> {
     protected String makeWktPoint(Double lon, Double lat) {
         return String.format("POINT(%s %s)", lon, lat);
     }
+
+    protected String makeWktLine(Double lon1, Double lat1, Double lon2, Double lat2) {
+        return String.format("LINESTRING(%s %s, %s %s)", lon1, lat1, lon2, lat2);
+    }
 }